[Feat] SunDrying, BluetoothShower, Swing Modifier
일광건조
이불두르기 방어력
기상 체력회복
두꺼운 이불 방어막

블루투스 샤워기
방수코팅 방어력
냉수마찰 이동속도
온수샤워 체력회복

휘두르기 세트
패링 탄환 제거

+++ Assets/02.Scripts/Skill/Skill/41010_BluetoothShower/BluetoothShowerModifier.cs
@@ -24,7 +24,7 @@ public class BluetoothShowerModifierData
 
     // 방수코팅
     public bool DefenseOnActive = false;
-    public int DefenseBonus = 0;
+    public float DefenseBonus = 0f;
 
     // 키친건!
     public bool AlwaysFire = false;
@@ -105,7 +105,7 @@ public class BluetoothKnockbackModifier : ActiveModifier<BluetoothShowerSkill>
 public class BluetoothDefenseModifier : ActiveModifier<BluetoothShowerSkill>
 {
     [Header("방어력 증가량")]
-    public int DefenseBonus = 1;
+    public float DefenseBonus = 1f;
 
     public override void ApplyModifier(BluetoothShowerSkill skill, ActiveUpgradeData upgradeData)
     {
